Extract a <Row /> component out of <Monitor />
